Added Rounds

+++ src/globals.h
@@ -13,11 +13,11 @@ extern const int gap_between_ships;
 extern const int title_width;
 extern const int title_height;
 
-extern const int title_top_left_x_offset;
-extern const int title_top_left_y_offset;
+extern const float title_top_left_x_offset;
+extern const float title_top_left_y_offset;
 
 extern const int start_button_width;
 extern const int start_button_height;
 
-extern const int start_button_top_left_x_offset;
-extern const int start_button_top_left_y_offset;
+extern const float start_button_top_left_x_offset;
+extern const float start_button_top_left_y_offset;
